src/backend/requirements.txt
python-dotenv==0.10.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
decorator==4.4.2
fastapi==0.95.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
multidict==6.0.4

openai==0.27.6
opencv-python==4.7.0.72
Pillow==9.5.0
proglog==0.1.10
psycopg2==2.9.6
pycparser==2.21
pydantic==1.10.7
PyJWT==2.7.0
PyMySQL==1.0.3
python-multipart==0.0.6
requests==2.30.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
yarl==1.9.2
